requirements.txt
streamlit==1.28.1
pandas==2.1.3
folium==0.14.0
streamlit-folium==0.13.0
psycopg2-binary==2.9.7
requests==2.31.0
python-dotenv==1.0.0 
altair==5.5.0
blinker==1.9.0
branca==0.8.1
cachetools==5.5.2
click==8.2.1
folium==0.14.0
gitdb==4.0.12
gitpython==3.1.44
importlib-metadata==6.11.0
markdown-it-py==3.0.0
mdurl==0.1.2
narwhals==1.46.0
numpy==1.24.3
packaging==23.2
pandas==2.1.3
pillow==10.4.0
plotly==5.17.0
protobuf==4.25.8
psycopg2-binary==2.9.7
pyarrow==20.0.0
pydeck==0.9.1
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
rich==13.9.4
smmap==5.0.2
streamlit==1.28.1
streamlit-folium==0.13.0
tenacity==8.5.0
toml==0.10.2
tzdata==2025.2
tzlocal==5.3.1
validators==0.35.0
watchdog==6.0.0
zipp==3.23.0